+++ migrate/sqlite.ts
@@ -1,0 +1,18 @@
+import { PlanetScaleDialect } from "kysely-planetscale";
+import { migrateToLatest } from "./migrate";
+
+const dialect = new PlanetScaleDialect({
+  host: process.env.DATABASE_HOST,
+  username: process.env.DATABASE_USERNAME,
+  password: process.env.DATABASE_PASSWORD,
+  fetch: (opts, init) =>
+    fetch(new Request(opts, { ...init, cache: undefined })),
+});
+
+migrateToLatest(dialect)
+  .then(() => {
+    console.log("migrated");
+  })
+  .catch((error) => {
+    console.error(error);
+  });
