+++ src/adapters/in-memory/clients/createClient.ts
@@ -1,7 +1,7 @@
-import { Client } from "../../../types";
+import { Client, PartialClient } from "../../../types";
 
-export function createClient(clientsStorage: Client[]) {
-  return async (client: Client): Promise<void> => {
+export function createClient(clientsStorage: PartialClient[]) {
+  return async (client: PartialClient): Promise<void> => {
     clientsStorage.push(client);
   };
 }
