+++ src/adapters/interfaces/Clients.ts
@@ -1,7 +1,7 @@
-import { Client } from "../../types";
+import { Client, PartialClient } from "../../types";
 
 export interface ClientsAdapter {
   get: (id: string) => Promise<Client | null>;
   // Temporary solution for testing
-  create?: (client: Client) => Promise<void>;
+  create?: (client: PartialClient) => Promise<void>;
 }
