+++ test/fixtures/context.ts
@@ -1,17 +1,9 @@
 // This is to make Request and other browser stuff work
 import "isomorphic-fetch";
 import { Context } from "hono";
-import {
-  AuthorizationResponseMode,
-  AuthorizationResponseType,
-  Client,
-  Env,
-  PasswordParams,
-  User,
-} from "../../src/types";
+import { Env, PartialClient, PasswordParams, User } from "../../src/types";
 import { oAuth2ClientFactory } from "./oauth2Client";
 import { mockedR2Bucket } from "./mocked-r2-bucket";
-import { kvStorageFixture } from "./kv-storage";
 import { EmailOptions } from "../../src/services/email/EmailOptions";
 import { Var } from "../../src/types/Var";
 import createAdapters from "../../src/adapters/in-memory";
@@ -20,6 +12,8 @@ import { sendLink, sendCode } from "../../src/controllers/email";
 import { Ticket } from "../../src/types/Ticket";
 import { OTP } from "../../src/types/OTP";
 import { Session } from "../../src/types/Session";
+import { client } from "./client";
+
 export interface ContextFixtureParams {
   headers?: { [key: string]: string };
   stateData?: { [key: string]: string };
@@ -28,7 +22,7 @@ export interface ContextFixtureParams {
   otps?: OTP[];
   passwords?: PasswordParams[];
   users?: User[];
-  clients?: KVNamespace;
+  clients?: PartialClient[];
   userData?: { [key: string]: string | boolean };
   email?: {
     sendLink?: typeof sendLink;
@@ -37,62 +31,10 @@ export interface ContextFixtureParams {
   logs?: any[];
 }
 
-interface stateInput {
-  state: string;
-  ttl?: number;
-}
-
-export const client: Client = {
-  id: "id",
-  name: "clientName",
-  client_secret: "clientSecret",
-  tenant_id: "tenantId",
-  allowed_callback_urls: ["http://localhost:3000", "https://example.com"],
-  allowed_logout_urls: ["http://localhost:3000", "https://example.com"],
-  allowed_web_origins: ["http://localhost:3000", "https://example.com"],
-  email_validation: "enabled",
-  tenant: {
-    sender_email: "senderEmail",
-    sender_name: "senderName",
-    audience: "audience",
-  },
-  connections: [
-    {
-      id: "connectionId1",
-      name: "google-oauth2",
-      client_id: "googleClientId",
-      client_secret: "googleClientSecret",
-      authorization_endpoint: "https://accounts.google.com/o/oauth2/v2/auth",
-      token_endpoint: "https://oauth2.googleapis.com/token",
-      response_mode: AuthorizationResponseMode.QUERY,
-      response_type: AuthorizationResponseType.CODE,
-      scope: "openid profile email",
-      created_at: "created_at",
-      updated_at: "updated_at",
-    },
-    {
-      id: "connectionId2",
-      name: "facebook",
-      client_id: "facebookClientId",
-      client_secret: "facebookClientSecret",
-      authorization_endpoint: "https://graph.facebook.com/oauth/access_token",
-      token_endpoint: "https://www.facebook.com/dialog/oauth",
-      response_mode: AuthorizationResponseMode.QUERY,
-      response_type: AuthorizationResponseType.CODE,
-      scope: "email public_profile",
-      created_at: "created_at",
-      updated_at: "updated_at",
-    },
-  ],
-  domains: [],
-};
-
 export function contextFixture(
   params?: ContextFixtureParams,
 ): Context<{ Bindings: Env; Variables: Var }> {
   const {
-    stateData = {},
-    userData = {},
     headers = {},
     logs = [],
     clients,
@@ -136,46 +78,20 @@ export function contextFixture(
     });
   }
 
+  if (clients) {
+    clients.forEach((client) => {
+      data.clients.create!(client);
+    });
+  } else {
+    data.clients.create!(client);
+  }
+
   // Add a known certificate
   data.certificates.upsertCertificates([getCertificate()]);
   // A test client
   if (!data.clients.create) {
     throw new Error("Missing create method on clients adapter");
   }
-  data.clients.create({
-    id: "clientId",
-    name: "Test Client",
-    connections: [
-      {
-        id: "connectionId1",
-        name: "google-oauth2",
-        client_id: "googleClientId",
-        client_secret: "googleClientSecret",
-        authorization_endpoint: "https://accounts.google.com/o/oauth2/v2/auth",
-        token_endpoint: "https://oauth2.googleapis.com/token",
-        response_mode: AuthorizationResponseMode.QUERY,
-        response_type: AuthorizationResponseType.CODE,
-        scope: "openid profile email",
-        created_at: "created_at",
-        updated_at: "updated_at",
-      },
-    ],
-    domains: [],
-    tenant_id: "tenantId",
-    allowed_callback_urls: [
-      "https://login.example.com/sv/callback",
-      "https://example.com",
-    ],
-    allowed_logout_urls: [],
-    allowed_web_origins: [],
-    email_validation: "enforced",
-    client_secret: "XjI8-WPndjtNHDu4ybXrD",
-    tenant: {
-      audience: "https://example.com",
-      sender_email: "login@example.com",
-      sender_name: "SenderName",
-    },
-  });
 
   return {
     set: () => {},
@@ -192,11 +108,6 @@ export function contextFixture(
       sendEmail: async (emailOptions: EmailOptions) => {
         logs.push(emailOptions);
       },
-      CLIENTS:
-        clients ||
-        kvStorageFixture({
-          clientId: JSON.stringify(client),
-        }),
       IMAGE_PROXY_URL: "https://imgproxy.dev.sesamy.cloud",
       data: {
         ...data,
